added dual thread ray marching

+++ cuda/includes/signed_distance.cu
@@ -154,7 +154,7 @@ __device__ float sd_mesh(const vec3 p, const unsigned int mesh_id, const SdRunti
     return sd;
 }
 
-__shared__ unsigned int composition_traversal_count[BLOCK_SIZE];
+__shared__ float composition_traversal_count[BLOCK_SIZE];
 
 __device__ float sd_composition(
     const vec3 p,
@@ -172,6 +172,8 @@ __device__ float sd_composition(
     BitSet<32> pos_mirrored {};
 
     RuntimeStackNode sd_runtime_stack[SD_RUNTIME_STACK_MAX_DEPTH];
+
+    // composition_traversal_count[threadIdx.x] += 1;
 
     vec3 position = p;
